FIX: Mudando o tipo do atributo da tabela Local

+++ backend/core/models/local.py
@@ -3,8 +3,8 @@ from .state import State
 
 
 class Local(models.Model):
-    longitude = models.BigIntegerField()
-    latitude = models.BigIntegerField()
+    longitude = models.DecimalField(decimal_places=2, max_digits=9)
+    latitude = models.DecimalField(decimal_places=2, max_digits=9)
     city = models.CharField(max_length=85)
     street_number = models.CharField(max_length=255)
     street_name = models.CharField(max_length=255)
